Data Validation , Filtering , Sorting , Pagination , JWT identity , User Registration

+++ FinsharkApi/Controllers/CommentController.cs
@@ -26,14 +26,19 @@ namespace FinsharkApi.Controllers
 
         public async Task<IActionResult> GetAllComments()
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
+            
             var comments = await _comRepo.GetAllComments();
             var commentDTOs = comments.Select(c => c.ToCommentDTO()).ToList();
             return Ok(commentDTOs);
         }
 
-        [HttpGet("get-comment/{id}")]
+        [HttpGet("get-comment/{id:int}")]
         public async Task<IActionResult> GetCommentById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var comment = await _comRepo.GetCommentById(id);
             if (comment == null)
             {
@@ -43,9 +48,11 @@ namespace FinsharkApi.Controllers
         }
 
 
-        [HttpPost("add-comment{stockId}")]
+        [HttpPost("add-comment{stockId:int}")]
         public async Task<IActionResult> AddComment([FromRoute] int stockId, CreateCommentDTO commentDTO)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             if (!await _stoRepo.StockExistsAsync(stockId))
             {
                 return BadRequest("Stock does not exist");
@@ -56,9 +63,11 @@ namespace FinsharkApi.Controllers
             return CreatedAtAction(nameof(GetCommentById), new { id = createdComment.Id }, createdComment.ToCommentDTO());
         }
 
-        [HttpPut("update-comment/{id}")]
+        [HttpPut("update-comment/{id:int}")]
         public async Task<IActionResult> UpdateComment([FromRoute] int id, UpdateCommentDTO commentDTO)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var comment = await _comRepo.UpdateComment(id, commentDTO.ToCommentUpdate());
             if (comment == null)
             {
@@ -67,9 +76,11 @@ namespace FinsharkApi.Controllers
             return Ok(comment.ToCommentDTO());
         }
 
-        [HttpDelete("delete-comment/{id}")]
+        [HttpDelete("delete-comment/{id:int}")]
         public async Task<IActionResult> DeleteComment([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var comment = await _comRepo.DeleteComment(id);
             if (comment == null)
             {
