+++ FinsharkApi/Controllers/StockController.cs
@@ -21,15 +21,17 @@ namespace FinsharkApi.Controllers
             _stockRepo = stockRepo;
         }
         [HttpGet("get-stocks")]
-        public async Task<IActionResult> GetStocks()
+        public async Task<IActionResult> GetStocks([FromQuery] Helpers.QueryObject query)
         {
-            var stocks = await _stockRepo.GetAllStocksAsync();
+            var stocks = await _stockRepo.GetAllStocksAsync(query);
             return Ok(stocks);
         }
 
-        [HttpGet("get-stock/{id}")]
+        [HttpGet("get-stock/{id:int}")]
         public async Task<IActionResult> GetStockById([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stock = await _stockRepo.GetStockByIdAsync(id);
             if (stock == null)
             {
@@ -42,6 +44,8 @@ namespace FinsharkApi.Controllers
         [Route("create-stock")]
         public async Task<IActionResult> Create([FromBody] CreateStockDTO stockDto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stockModel = stockDto.ToStockFromCreateDTO();
               await _stockRepo.CreateStockAsync(stockModel);
                 return CreatedAtAction(nameof(GetStockById), new { id = stockModel.Id }, stockModel);
@@ -50,9 +54,11 @@ namespace FinsharkApi.Controllers
         }
 
         [HttpPut]
-        [Route("update-stock/{id}")]
+        [Route("update-stock/{id:int}")]
         public async Task<IActionResult> Update([FromRoute] int id, [FromBody] UpdateStockDTO dto)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stock = await _stockRepo.UpdateStockAsync(id, dto);
             if (stock == null)
             {
@@ -62,10 +68,12 @@ namespace FinsharkApi.Controllers
         }
 
         [HttpDelete]
-        [Route("delete-stock/{id}")]
+        [Route("delete-stock/{id:int}")]
 
         public async Task<IActionResult> DeleteStock([FromRoute] int id)
         {
+            if (!ModelState.IsValid)
+                return BadRequest(ModelState);
             var stock = await _stockRepo.DeleteStockAsync(id);
             if (stock == null)
             {
